Handle genesis block

+++ JSON.go
@@ -300,7 +300,7 @@ type RawTransaction struct {
 
 // Vout represent an OUT value
 type Vout struct {
-	Value        float64      `json:"value"`
+	Value        uint64       `json:"value"`
 	N            int          `json:"n"`
 	ScriptPubKey ScriptPubKey `json:"scriptPubKey"`
 }
@@ -309,7 +309,7 @@ type Vout struct {
 type Vin struct {
 	Coinbase  string    `json:"coinbase"`
 	Txid      string    `json:"txid"`
-	Vout      int       `json:"vout"`
+	Vout      uint64    `json:"vout"`
 	ScriptSig ScriptSig `json:"scriptSig"`
 	Sequence  uint32    `json:"sequence"`
 }
